requirements.txt
Flask
openai
flask-cors
python-dotenv